DataSpec: Avoid unnecessary copying
The key of a map is actually const, so a pair where the first element is
not const will actually require a copy of the data to be made, given the
assigned type wouldn't be equivalent.

There was also implicit conversion to DNAFourCC where it wasn't
necessary. Instead, be explicit about the construction of the type and
calling write(). This way we only construct the new DNAFourCC type and
not the rest of the contents of the pair.

While we're at it, we can make use of structured bindings.